refact: separate out media model
This commit separates out the media model. Additionally, it creates a
common place in the file structure to place future models, if needed.
